dockerfile: omit inadvertant component upgrades
An issue was reported as part of CRI-O pipeline testsing,
where a class of systems (specifically RHEL 7) were  failing.

While we stuck to `quay.io/buildah/stable:v1.9.0` ,
the `yum update` construct in the Dockerfile was
upgrading the installed components in the image,
including `buildah`, taking us back to square one.

All what we want is just `wget`, so skip `yum update`.