dockerfile: upgrade buildah
buildah was fixed at v1.9.0 due to bind mount issues
when interacted with a buggy fuse-overlay module.
But this proved to be poorly performing for certain
stacks when exercised in the pipeline.

Upgrade buildah to latest that has the fix for the performance
issue, while documenting the requirement for fuse-overlay

Fixes: https://github.com/kabanero-io/kabanero-pipelines/issues/121
Refs: https://github.com/containers/buildah/issues/2047